requirements.txt
Office365-REST-Python-Client